fix: Collision onesided
The collision in mujoco is displayed as two arrays, where each collision pair share the same index. So far the check was only one sided. Meaning that if collision(geom1, geom2) was used, it was only checked if geom1 is in array1 and geom2 is in array2, not the other way around.